Extract run package
