+++ exo-go/src/dockercompose/create_new_container.go
@@ -1,8 +1,8 @@
 package dockercompose
 
-import "github.com/Originate/exosphere/exo-go/src/util"
+import "github.com/Originate/exosphere/exo-go/src/run"
 
 // CreateNewContainer creates a new docker container for the given service
 func CreateNewContainer(serviceName string, env []string, dockerComposeDir string, logChannel chan string) error {
-	return util.RunAndLog(dockerComposeDir, env, logChannel, "docker-compose", "create", "--build", serviceName)
+	return run.AndLog(dockerComposeDir, env, logChannel, "docker-compose", "create", "--build", serviceName)
 }
