+++ exo-go/src/dockercompose/pull_all_images.go
@@ -1,8 +1,8 @@
 package dockercompose
 
-import "github.com/Originate/exosphere/exo-go/src/util"
+import "github.com/Originate/exosphere/exo-go/src/run"
 
 // PullAllImages pulls all the docker images defined in docker-compose.yml
 func PullAllImages(dockerComposeDir string, logChannel chan string) error {
-	return util.RunAndLog(dockerComposeDir, []string{}, logChannel, "docker-compose", "pull")
+	return run.AndLog(dockerComposeDir, []string{}, logChannel, "docker-compose", "pull")
 }
