+++ exo-go/src/dockercompose/restart_container.go
@@ -1,8 +1,8 @@
 package dockercompose
 
-import "github.com/Originate/exosphere/exo-go/src/util"
+import "github.com/Originate/exosphere/exo-go/src/run"
 
 // RestartContainer starts the docker container of the given service
 func RestartContainer(serviceName string, env []string, dockerComposeDir string, logChannel chan string) error {
-	return util.RunAndLog(dockerComposeDir, env, logChannel, "docker-compose", "restart", serviceName)
+	return run.AndLog(dockerComposeDir, env, logChannel, "docker-compose", "restart", serviceName)
 }
