+++ exo-go/src/template/index.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"path"
 
+	"github.com/Originate/exosphere/exo-go/src/run"
 	"github.com/Originate/exosphere/exo-go/src/util"
 	"github.com/pkg/errors"
 	"github.com/tmrts/boilr/pkg/template"
@@ -16,7 +17,7 @@ const templatesDir = ".exosphere"
 // Add fetches a remote template from GitHub and stores it
 // under templateDir, returns an error if any
 func Add(gitURL, templateName, templateDir string) error {
-	if output, err := util.Run("", "git", "submodule", "add", "--name", templateName, gitURL, templateDir); err != nil {
+	if output, err := run.Run("", "git", "submodule", "add", "--name", templateName, gitURL, templateDir); err != nil {
 		return errors.Wrap(err, fmt.Sprintf("Failed to fetch template from %s: %s\n", gitURL, output))
 	}
 	return nil
@@ -43,7 +44,7 @@ func CreateTmpServiceDir(appDir, chosenTemplate string) (string, error) {
 
 // Fetch fetches updates for all existing remote templates
 func Fetch() error {
-	if output, err := util.Run("", "git", "submodule", "foreach", "git", "pull", "origin", "master"); err != nil {
+	if output, err := run.Run("", "git", "submodule", "foreach", "git", "pull", "origin", "master"); err != nil {
 		return errors.Wrap(err, fmt.Sprintf("Failed to fetch updates for existing templates: %s\n", output))
 	}
 	return nil
@@ -89,7 +90,7 @@ func Remove(templateName, templateDir string) error {
 	denitCommand := []string{"git", "submodule", "deinit", "-f", templateDir}
 	removeModulesCommand := []string{"rm", "-rf", fmt.Sprintf(".git/modules/%s", templateName)}
 	gitRemoveCommand := []string{"git", "rm", "-f", templateDir}
-	return util.RunSeries("", [][]string{denitCommand, removeModulesCommand, gitRemoveCommand})
+	return run.Series("", [][]string{denitCommand, removeModulesCommand, gitRemoveCommand})
 }
 
 // Helpers
